cloud_function/requirements.txt
google-api-python-client==2.86.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.36.0
google-cloud-pubsub==2.18.0
google-cloud-secretmanager==2.16.0
requests==2.31.0
python-dateutil==2.8.2
